+++ ParallaxGen/src/main.cpp
@@ -207,12 +207,6 @@ void mainRunner(ParallaxGenCLIArgs &Args, const filesystem::path &ExePath) {
     exit(1);
   }
 
-  if (find(ActivePlugins.begin(), ActivePlugins.end(), L"simplicity of snow.esp") != ActivePlugins.end() &&
-      (Params.ShaderPatcher.TruePBR || Params.ShaderPatcher.ComplexMaterial)) {
-    spdlog::warn("You have Simplicity of Snow installed. SoS is incompatible with complex material and PBR. Use a "
-                 "single-pass snow mod such as Better Dynamic Snow v3 instead.");
-  }
-
   // Init PGP library
   if (Params.Processing.PluginPatching) {
     spdlog::info("Initializing plugin patcher");